Move bmi module to _bmi
